requirements.txt
tox>=2.1.1
mock>=1.3.0
coverage>=3.7.1
django-nose>=1.4
django-braces>=1.8.0
freezegun>=0.3.5
nose-parameterized>=0.5.0

django-import-export==0.5.1